+++ run.sh
@@ -1,2 +1,2 @@
-source install/setup.sh
+source ./install/setup.sh
 ros2 launch autonomousr_explorer_bringup full_system.launch.py